requirements.txt
# To build: run in terminal
# ./build_env.sh
torch==2.3.0
transformers==4.57.0
#transformers==4.52.3
jetson-utils==0.5.1
numpy==1.26.4
Pillow==10.4.0
pygame==2.5.2
argparse